+++ build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
 }
 
-group 'org.example'
+group 'com.fastcampus'
 version '1.0-SNAPSHOT'
 
 repositories {
@@ -10,6 +10,9 @@ repositories {
 }
 
 dependencies {
+    implementation 'org.springframework:spring-context:5.3.15'
+
+    testImplementation 'org.springframework:spring-test:5.3.15'
     testImplementation 'org.assertj:assertj-core:3.22.0'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
