+++ src/main/java/com/fastcampus/practice/config/Config.java
@@ -1,0 +1,10 @@
+package com.fastcampus.practice.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+
+@ComponentScan("com.fastcampus.practice")
+@Configuration
+public class Config {
+}
